refactor(emitter): split module emission analysis
